fix: sonarqube duplicates
