+++ src/components/CommunityPortal/CPDashboard.jsx
@@ -13,6 +13,7 @@ import styles from './CPDashboard.module.css';
 import { ENDPOINTS } from '../../utils/URL';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
+import { isTomorrow, isComingWeekend } from './utils';
 
 const RECENT_SEARCHES_KEY = 'cp_recent_searches';
 const MAX_RECENT_SEARCHES = 10;
@@ -127,29 +128,6 @@ const DEFAULT_FILTERS = {
   categories: '',
 };
 
-function isTomorrow(dateString) {
-  const input = new Date(dateString);
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  const tomorrow = new Date(today);
-  tomorrow.setDate(today.getDate() + 1);
-  return input >= tomorrow && input < new Date(tomorrow.getTime() + 24 * 60 * 60 * 1000);
-}
-
-function isComingWeekend(dateString) {
-  const input = new Date(dateString);
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  const day = today.getDay();
-  const daysUntilSaturday = (6 - day + 7) % 7 || 7;
-  const saturday = new Date(today);
-  saturday.setDate(today.getDate() + daysUntilSaturday);
-  const sunday = new Date(saturday);
-  sunday.setDate(saturday.getDate() + 1);
-  sunday.setHours(23, 59, 59, 999);
-  return input >= saturday && input <= sunday;
-}
-
 function passesFilters(
   event,
   { showPastEvents, isPastEvent, onlineOnly, dateFilter, searchQuery },
@@ -242,7 +220,7 @@ export function CPDashboard() {
   };
 
   // keep this near your refs/functions
-  const BASE_HEIGHT = 36;
+  const BASE_HEIGHT = 32;
 
   const autoGrow = el => {
     if (!el) return;
